+++ .gitignore
@@ -90,4 +90,4 @@ TensorRT-LLM
 # Local build artifacts for devcontainer
 .build/
 # Copied binaries to ignore
-deploy/dynamo/sdk/src/dynamo/sdk/cli/bin
+deploy/sdk/src/dynamo/sdk/cli/bin
