+++ container/Dockerfile.vllm
@@ -302,7 +302,7 @@ COPY hatch_build.py /workspace/
 COPY lib/ /workspace/lib/
 COPY components /workspace/components
 COPY launch /workspace/launch
-COPY deploy/dynamo/sdk /workspace/deploy/dynamo/sdk
+COPY deploy/sdk /workspace/deploy/sdk
 
 # Build Rust crate binaries packaged with the wheel
 RUN cargo build --release --locked --features mistralrs,sglang,vllm,python \
@@ -313,15 +313,15 @@ RUN cargo build --release --locked --features mistralrs,sglang,vllm,python \
     -p metrics
 
 # Create symlinks for wheel building
-RUN mkdir -p /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/ && \
+RUN mkdir -p /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/ && \
     # Remove existing symlinks
-    rm -f /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/* && \
+    rm -f /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/* && \
     # Create new symlinks pointing to the correct location
-    ln -sf /workspace/target/release/dynamo-run /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/dynamo-run && \
-    ln -sf /workspace/target/release/http /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/http && \
-    ln -sf /workspace/target/release/llmctl /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/llmctl && \
-    ln -sf /workspace/target/release/metrics /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/metrics && \
-    ln -sf /workspace/target/release/mock_worker /workspace/deploy/dynamo/sdk/src/dynamo/sdk/cli/bin/mock_worker
+    ln -sf /workspace/target/release/dynamo-run /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/dynamo-run && \
+    ln -sf /workspace/target/release/http /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/http && \
+    ln -sf /workspace/target/release/llmctl /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/llmctl && \
+    ln -sf /workspace/target/release/metrics /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/metrics && \
+    ln -sf /workspace/target/release/mock_worker /workspace/deploy/sdk/src/dynamo/sdk/cli/bin/mock_worker
 
 # Build dynamo wheel
 RUN uv build --wheel --out-dir /workspace/dist && \
