+++ deploy/cloud/operator/PROJECT
@@ -6,7 +6,7 @@ domain: nvidia.com
 layout:
 - go.kubebuilder.io/v4
 projectName: dynamo-kubernetes-operator
-repo: github.com/ai-dynamo/dynamo/deploy/dynamo/operator/api/v1alpha1
+repo: github.com/ai-dynamo/dynamo/deploy/cloud/operator/api/v1alpha1
 resources:
 - api:
     crdVersion: v1
@@ -14,14 +14,14 @@ resources:
   controller: true
   domain: nvidia.com
   kind: DynamoComponentDeployment
-  path: github.com/ai-dynamo/dynamo/deploy/dynamo/operator/api/v1alpha1
+  path: github.com/ai-dynamo/dynamo/deploy/cloud/operator/api/v1alpha1
   version: v1alpha1
 - api:
     crdVersion: v1
     namespaced: true
   domain: nvidia.com
   kind: DynamoComponent
-  path: github.com/ai-dynamo/dynamo/deploy/dynamo/operator/api/v1alpha1
+  path: github.com/ai-dynamo/dynamo/deploy/cloud/operator/api/v1alpha1
   version: v1alpha1
 - api:
     crdVersion: v1
@@ -29,6 +29,6 @@ resources:
   controller: true
   domain: nvidia.com
   kind: DynamoGraphDeployment
-  path: github.com/ai-dynamo/dynamo/deploy/dynamo/operator/api/v1alpha1
+  path: github.com/ai-dynamo/dynamo/deploy/cloud/operator/api/v1alpha1
   version: v1alpha1
 version: "3"
