+++ docs/guides/dynamo_deploy/dynamo_cloud.md
@@ -138,7 +138,7 @@ export PIPELINES_DOCKER_PASSWORD=<your-docker-password>
 2. [One-time Action] Create a new kubernetes namespace and set it as your default.
 
 ```bash
-cd deploy/dynamo/helm
+cd deploy/cloud/helm
 kubectl create namespace $NAMESPACE
 kubectl config set-context --current --namespace=$NAMESPACE
 ```
