+++ pyproject.toml
@@ -78,7 +78,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["deploy/dynamo/sdk/src/dynamo", "components/planner/src/dynamo"]
+packages = ["deploy/sdk/src/dynamo", "components/planner/src/dynamo"]
 
 # This section is for including the binaries in the wheel package
 # but doesn't make them executable scripts in the venv bin directory
@@ -132,7 +132,7 @@ addopts = [
     "--mypy",
     "--ignore-glob=*model.py",
     "--ignore-glob=*_inc.py",
-    "--ignore-glob=deploy/dynamo/api-store/*",
+    "--ignore-glob=deploy/cloud/api-store/*",
     # FIXME: Get relative/generic blob paths to work here
 ]
 xfail_strict = true
